add secureRoutes Component
